add CJumpRangeModel to detect bottom and top height of each bird jump auto learn best bottom height offset which makes worst possibility to crash bottom and top pipe

+++ FlappyBirdAI/util/CPerformanceCounter.h
@@ -41,7 +41,7 @@ private:
 
 
     const char* m_szName ;
-    LONGLONG m_tmpCounter ;
+    double m_tmpCounter ;
 
 };
